# openapi.yaml
openapi: 3.0.3
info:
  title: API de Consulta de Precios
  version: 1.0.0
  description: API para consultar precios aplicables a productos

servers:
  - url: http://localhost:8080
    description: Servidor local para desarrollo
  - url: https://api-pricing-service-936595159798.us-central1.run.app
    description: Servidor de producción

paths:
  /prices:
    get:
      tags:
        - PricingController
      operationId: getApplicablePrice
      summary: Consultar precio aplicable por producto, marca y fecha
      parameters:
        - in: query
          name: brandId
          required: true
          schema:
            type: integer
            format: int32
          description: Identificador de la cadena (brand)
        - in: query
          name: productId
          required: true
          schema:
            type: integer
            format: int32
          description: Identificador del producto
        - in: query
          name: applicationDate
          required: true
          schema:
            type: string
            format: date-time
          description: Fecha de aplicación en formato ISO-8601
      responses:
        '200':
          description: Precio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        '404':
          description: No se encontró precio aplicable

  /products/{productId}/prices:
    get:
      tags:
        - PricingController
      operationId: getAllPricesByProductId
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
            default: 35456
            example: 35456
      summary: Obtener todos los precios por producto
      responses:
        '200':
          description: Lista de precios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceResponse'

  /prices/{id}:
    get:
      tags:
        - PricingController
      operationId: getPriceById
      summary: Obtener el precio del producto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            default: 1
            example: 1
      responses:
        '200':
          description: Precio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        '404':
          description: Precio no encontrado
components:
  schemas:
    PriceResponse:
      type: object
      properties:
        productId:
          type: integer
        brandId:
          type: integer
        priceList:
          type: integer
        price:
          type: number
          format: float
        appDate:
          type: string
          format: date-time
        currency:
          type: string